Two Crystal balls(Binary search of O(sqrt(n)))
The input array would be like this:
[false, false, false, true, true, true]
The function will goes each steps of sqrt(n). First it will check
whether the value is true or not. If the value goes true it means we
recived to the break point of crystals. Then the algorithm start to
search values between each steps' block.
The complexity of this algorithm is O(sqrt(n)) which is a little bit
slower than the original O(n) algorithm. But in this case it will act
more efficient.