Stack: set pop()
First it will handles the length(The max method will keep us safe from
negative values. It would be like this.length-- but it could be -1 in a
time so the math(0, this.length -1) will keep the value at least 0).
Then handle very first node and other places nodes too.